README.md
# Tour Rwanda
#### A site for adventurers in Rwanda
#### By Kobusingye Naume.
## Description
This is a site for adventurers in Rwanda where people will go and checkout the tourist attractions in Rwanda.
## Setup/Installation Requirements
* folk it 
* clone it from the github account
* download and edit it on your computer
## Known Bugs
there are no issues i know of for now.
## Technologies Used
I used html, css, and java script and jquerry.
## Support and contact details
Feel free to reach out to me when you happen to have any issue or anything else related to this project. contact me on my email at kobnaome@gmail.com.
## BDD
### behaviour
* select if you want  to see the museums.
* select if you want to see the  water bodies.
* select if you want to see the hotels and accomodations.
* select if you want to see the forests and National parks.
* add more orders or just shop.
### input examples
* select if you want to book accomodations
* hover over the images on the home  to see the magic
### output example
* on mouse hover shows the name of the place in the image .
### Link to my repository.
https://naume-k.github.io/Delani-studio/
### License
Copyright © 2019 Tour-Rwanda.